requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
environs==9.3.1
Flask==1.1.2
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gspread==3.6.0
httplib2==0.19.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.10.0
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
